pyproject.toml
[tool.poetry]
name = "pi-ai-os"
version = "0.0.4"
description = "Integrate your OS with an LLM."
authors = ["piEsposito <piero.skywalker@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
tiny-ai-client = "^0.0.11"
pillow = "^10.4.0"
omegaconf = "^2.3.0"

[tool.poetry.scripts]
aios = "pi_ai_os:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
